updated the score mechanic to now delete any ball that crosses the bounds and generate a new one. this will allow for mini balls (created from splitting a ball when its hit with a laser) to exist temporarily until they cross the bounds then a regular ball replaces them.
running into an issue with the object factory delete method. there is an overlap occurring with object ID's during rapid / multiple deletion events
because lasers have to be deleted after a collision or crossing a bound and balls need to be deleted when crossing the bounds. when these two events
overlap there is a bug with the deletion where you can end up with no balls on the board.